[김덕인] 운영자 강사 이미지 추가 (db에 추가 및 resource에 파일 자동 추가)

+++ lms/src/main/java/gd/fintech/lms/vo/MypageImage.java
@@ -7,7 +7,7 @@ public class MypageImage {
 	private String mypageImageUuid;			// 마이페이지 이미지 UUID
 	private String mypageImageOriginal;		// 마이페이지 이미지 원본 이름
 	private String accountId;				// 마이페이지 계정 아이디
-	private int mypageImageSize;			// 이미지 크기
+	private long mypageImageSize;			// 이미지 크기
 	private String mypageImageType;			// 이미지 타입
 	private String mypageImageCreateDate;	// 이미지 생성 일시
 }
